+++ NewNew/SB_UP/USER/Task/Yaw_task.h
@@ -7,9 +7,9 @@ void Yaw_Task(void * argument);
 
 static void Yaw_init(void);
 extern float motor_value(int16_t k, int16_t n);
-extern void yaw_control_L(int16_t max_angle, int16_t min_angle);
-extern void yaw_control_R(int16_t max_angle, int16_t min_angle);
-extern void yaw_finding_L(int16_t max_angle, int16_t min_angle);
-extern void yaw_finding_R(int16_t max_angle, int16_t min_angle);
+extern void yaw_control_L(float max_angle, float min_angle);
+extern void yaw_control_R(float max_angle, float min_angle);
+extern void yaw_finding_L(float max_angle, float min_angle);
+extern void yaw_finding_R(float max_angle, float min_angle);
 
 #endif
